+++ MovieDbAssistant.App/rsc/html/templates/dark/pages/page-details.tpl.html
