+++ MovieDbAssistant.App/rsc/html/templates/dark/pages/page-list.tpl.html
@@ -1,14 +1,5 @@
 ﻿<!DOCTYPE html>
-<html lang="{{lang}}"
-      data-template-id="{{templateId}}"
-      data-template-version="{{templateVersion}}"
-      data-template-version-date="{{templateVersionDate}}"
-      data-software="{{software}}"
-      data-software-id="{{softwareId}}"
-      data-software-version="{{softwareVersion}}"
-      data-software-version-date="{{softwareVersionDate}}"
-      data-built-at="{{builtAt}}"
-      >
+<html {{{html}}}>
 <head>
     <title>{{pageTitleList}} | {{subTitleList}}</title>
     <meta charset="utf-8" />
