+++ MovieDbAssistant.App/rsc/html/templates/dark/parts/html.tpl.html
@@ -1,0 +1,9 @@
+﻿lang="{{lang}}"
+data-template-id="{{templateId}}"
+data-template-version="{{templateVersion}}"
+data-template-version-date="{{templateVersionDate}}"
+data-software="{{software}}"
+data-software-id="{{softwareId}}"
+data-software-version="{{softwareVersion}}"
+data-software-version-date="{{softwareVersionDate}}"
+data-built-at="{{builtAt}}"
